refactor: Create exchange-rate-info/
