+++ src/app/app-routing.module.ts
@@ -5,7 +5,7 @@ import { PageNotFoundComponent } from './page-not-found/page-not-found.component
 const routes: Routes = [
     {
         path: 'converter',
-        loadComponent: () => import('./converter/converter.component').then(m => m.ConverterComponent),
+        loadComponent: () => import('./exchange-rate-info/converter/converter.component').then(m => m.ConverterComponent),
     },
     {
         path: '',
