+++ src/app/exchange-rate-info/converter/converter.component.spec.ts
@@ -4,7 +4,7 @@ import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ExchangeRateService } from 'src/app/services';
 import { iso4217 } from 'src/app/exchange-rate/iso4217.enum';
-import { ExchangeRateMockInterceptor } from '../exchange-rate/exchange-rate-mock.interceptor';
+import { ExchangeRateMockInterceptor } from '../../exchange-rate/exchange-rate-mock.interceptor';
 
 describe('ConverterComponent', () => {
     let component: ConverterComponent;
