+++ src/app/exchange-rate-info/exchange-rate-card/exchange-rate-card.component.scss
