+++ src/app/exchange-rate-info/exchange-rate-infos.type.ts
